Fixed commit comments
style, indentation and comments fixed
