+++ renta_facil_colima/src/hooks/ExtraInfoState.js
@@ -3,7 +3,9 @@ import * as Yup from 'yup';
 const createExtraInfoSchema = () => {
   return Yup.object().shape({
     address: Yup.string().required('Address field can not be empty'),
-    profilePicture: Yup.string().notRequired('Profile Picture field can not be empty'),
+    profilePicture: Yup.string().required(
+      'Profile Picture field can not be empty',
+    ),
     birthday: Yup.date().required('Birthday field can not be empty'),
   });
 };
